--- backend/src/main/java/dev/tripdraw/common/dto/SearchPaging.java
@@ -1,9 +1,0 @@
-package dev.tripdraw.common.dto;
-
-import dev.tripdraw.common.domain.Paging;
-
-public record SearchPaging(Long lastViewedId, Integer limit) {
-    public Paging toPaging() {
-        return new Paging(lastViewedId, limit);
-    }
-}
