+++ backend/src/main/java/dev/tripdraw/post/application/query/PostQueryService.java
@@ -1,0 +1,5 @@
+package dev.tripdraw.post.application.query;
+
+public class PostQueryService {
+
+}
