--- backend/src/main/java/dev/tripdraw/post/domain/PostCustomRepository.java
@@ -1,10 +1,0 @@
-package dev.tripdraw.post.domain;
-
-import dev.tripdraw.common.domain.Paging;
-
-import java.util.List;
-
-public interface PostCustomRepository {
-
-    List<Post> findAllByConditions(SearchConditions conditions, Paging paging);
-}
