+++ backend/src/main/java/dev/tripdraw/post/dto/query/PostSearchPaging.java
@@ -1,0 +1,4 @@
+package dev.tripdraw.post.dto.query;
+
+public record PostSearchPaging(Long lastViewedId, Integer limit) {
+}
